Move Lab2 to its own folder
