+++ polyanskiyIlya/src/views/lab2/Lab2.tsx
